Fix multiple varargs-related issues (#SCL-12708, #SCL-14152) fixed
Previously being 'repeated' was a characteristic of an ScParameter,
which is actually incorrect and has led to many inconsistencies (i.e.
def f(x: Int*): Unit = ???; x.`type`() was Int and not Seq[Int] which
is almost always undesirable).

It also resulted in an ugly pattern, where one would have to explicitly
match on an ScParameter to call param.getRealParameterType, which
in turn effectively made ScParameter !<: Typeable. Forgetting to do so
resulted in bugs (e.g. in OverridingAnnotator).

This introduces ScTypeElement#isRepeated method, which will help handle
cases where repeated type element is not in an immediate parameter
position (e.g. def f(creator: (Int*) => Long): Long = ???).

Compatibility.scala is refactored to support this inversion of old
logic.